# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API с JWT авторизацией",
        "title": "Fakegram API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/messages/private-chat/{user_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает сообщения из приватного чата с поддержкой двунаправленной пагинации.\nПри первой загрузке (direction=around) автоматически позиционирует на первом непрочитанном сообщении.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "messages"
                ],
                "summary": "Получить сообщения чата",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID второго пользователя",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "around",
                            "older",
                            "newer"
                        ],
                        "type": "string",
                        "default": "around",
                        "description": "Направление загрузки: around (первая загрузка вокруг непрочитанного), older (старые сообщения), newer (новые сообщения)",
                        "name": "direction",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Временная метка опорного сообщения в формате RFC3339 (например, 2024-01-02T15:04:05Z). Обязателен для direction=older или direction=newer",
                        "name": "cursor",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "type": "integer",
                        "default": 30,
                        "description": "Количество сообщений (1-100)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список сообщений с информацией о пагинации",
                        "schema": {
                            "$ref": "#/definitions/models.GetMessagesResponse"
                        }
                    },
                    "400": {
                        "description": "Неверные параметры (неверный direction, неверный формат cursor и т.д.)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен (пользователь не является участником чата)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/messages/send": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Отправляет сообщение. Для существующих диалогов используйте chat_id, для новых - receiver_id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "messages"
                ],
                "summary": "Отправить сообщение",
                "parameters": [
                    {
                        "description": "Данные для отправки сообщения",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Сообщение отправлено",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Неверные данные запроса",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизован",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/messages/{message_id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Редактирование текста сообщения. Только отправитель может редактировать свое сообщение",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "messages"
                ],
                "summary": "Редактировать сообщение",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID сообщения для редактирования",
                        "name": "message_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные для редактирования сообщения",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Сообщение успешно отредактировано",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Неверные параметры запроса",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизован",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Сообщение не найдено",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Удаление сообщения. Только отправитель может удалить свое сообщение",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "messages"
                ],
                "summary": "Удалить сообщение",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID сообщения для удаления",
                        "name": "message_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Сообщение успешно удалено",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Неверные параметры запроса",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизован",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Сообщение не найдено",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.CreateMessageRequest": {
            "type": "object",
            "required": [
                "chat_id",
                "message_text",
                "message_type"
            ],
            "properties": {
                "chat_id": {
                    "type": "string"
                },
                "message_text": {
                    "type": "string"
                },
                "message_type": {
                    "enum": [
                        "text",
                        "image",
                        "file",
                        "video",
                        "audio"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.MessageType"
                        }
                    ]
                },
                "receiver_id": {
                    "type": "string"
                },
                "reply_to_message_id": {
                    "type": "string"
                }
            }
        },
        "models.GetMessagesResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "cursors": {
                    "$ref": "#/definitions/models.MessageCursors"
                },
                "first_msg_time": {
                    "type": "string"
                },
                "has_more_newer": {
                    "type": "boolean"
                },
                "has_more_older": {
                    "type": "boolean"
                },
                "last_msg_time": {
                    "type": "string"
                },
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.MessageDetail"
                    }
                },
                "total_unread": {
                    "type": "integer"
                }
            }
        },
...
        "models.MessageCursors": {
            "type": "object",
            "properties": {
                "newer": {
                    "type": "string"
                },
                "older": {
                    "type": "string"
                }
            }
        },
        "models.MessageDetail": {
            "type": "object",
            "required": [
                "sender_id"
            ],
            "properties": {
                "chat_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_deleted": {
                    "type": "boolean"
                },
                "is_edited": {
                    "type": "boolean"
                },
                "is_read": {
                    "type": "boolean"
                },
                "message_text": {
                    "type": "string"
                },
                "message_type": {
                    "$ref": "#/definitions/models.MessageType"
                },
                "read_at": {
                    "type": "string"
                },
                "reply_to_message": {
                    "$ref": "#/definitions/models.MessageDetail"
                },
                "reply_to_message_id": {
                    "type": "string"
                },
                "sender": {
                    "$ref": "#/definitions/models.UserDetail"
                },
                "sender_id": {
                    "type": "string"
                },
                "update_at": {
                    "type": "string"
                }
            }
        },
        "models.MessageType": {
            "type": "string",
            "enum": [
                "text",
                "image",
                "file"
            ],
            "x-enum-varnames": [
                "MessageTypeText",
                "MessageTypeImage",
                "MessageTypeFile"
            ]
        },
...
        "models.UpdateMessageRequest": {
            "type": "object",
            "required": [
                "message_text"
            ],
            "properties": {
                "message_text": {
                    "type": "string"
                }
            }
        },
        "models.UserDetail": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}